RAD-22 - Web related code mixed with api
move web related code (controllers, extension) to omod
