+++ omod/src/main/java/org/openmrs/module/radiology/web/controller/PortletsController.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology.web;
+package org.openmrs.module.radiology.web.controller;
 
 import java.text.ParseException;
 import java.util.Date;
