+++ omod/src/main/java/org/openmrs/module/radiology/web/controller/StaticPagesController.java
@@ -7,26 +7,25 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology.web;
+package org.openmrs.module.radiology.web.controller;
 
-import org.openmrs.api.context.Context;
-import org.openmrs.module.radiology.RadiologyActivator;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 @Controller
-@RequestMapping("/module/radiology/radiologyOrder.list")
-public class RadiologyOrderListController {
+public class StaticPagesController {
 	
-	@RequestMapping(method = RequestMethod.GET)
-	public ModelAndView handleRequest() {
-		ModelAndView mav = new ModelAndView();
-		mav.setViewName("module/radiology/radiologyOrderList");
-		if (RadiologyActivator.badInit(Context.getUserService(), Context.getOrderService())) {
-			mav.addObject("initialized", "fail");
-		}
+	/**
+	 * Called resources because returns a path /module/radiology/resources/+path
+	 * Example path /module/radiology/static/felix-config.list
+	 * @param path extracted from URL /module/radiology/static/xxx.list 
+	 * @return ModelAndView with the view pointing to the file /web/module/resources/xxx.jsp
+	 */
+	@RequestMapping("/module/radiology/static/{path}")
+	ModelAndView resources(@PathVariable("path") String path) {
+		ModelAndView mav = new ModelAndView("/module/radiology/resources/" + path);
 		return mav;
 	}
 }
