+++ omod/src/main/java/org/openmrs/module/radiology/web/extension/html/RadiologyDashboardExt.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology.extension.html;
+package org.openmrs.module.radiology.web.extension.html;
 
 import org.openmrs.module.Extension;
 import org.openmrs.module.web.extension.PatientDashboardTabExt;
