+++ omod/src/main/resources/config.xml
@@ -24,12 +24,12 @@
 	<!-- Extensions -->
 	<extension>
 		<point>org.openmrs.admin.list</point>
-		<class>@MODULE_PACKAGE@.extension.html.AdminList</class>
+		<class>@MODULE_PACKAGE@.web.extension.html.AdminList</class>
 	</extension>
 
 	<extension>
 		<point>org.openmrs.patientDashboardTab</point>
-		<class>@MODULE_PACKAGE@.extension.html.RadiologyDashboardExt</class>
+		<class>@MODULE_PACKAGE@.web.extension.html.RadiologyDashboardExt</class>
 	</extension>
 
 	<!-- /Extensions -->
